Database 1

+++ TaxiAppWebApi/Models/Analytics.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public int TotalRides { get; set; }
         public int TotalUsers { get; set; }
-        public decimal TotalRevenue { get; set; }
+        public float TotalRevenue { get; set; }
         public DateTime GeneratedAt { get; set; }
     }
